@@ -3,11 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.2.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "@rmwc/top-app-bar": "^5.6.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-router-dom": "^5.0.1",
-    "react-scripts": "^2.1.8",
+    "react-scripts": "^3.1.1",
+    "reflexbox": "^3.0.1",
     "styled-components": "^4.3.2"
   },
   "scripts": {
